app_code/requirements.txt
aggdraw==1.3.12
appnope==0.1.2
h5py==3.6.0
numpy==1.21.4
Pillow==8.4.0
PySDL2==0.9.11
psutil==5.9.0
pysdl2-dll==2.0.16
pyusb==1.2.1 
sounddevice==0.4.3
SoundFile==0.10.3.post1